v0.0.12 - Add badges - Use constant during test instead of rewrite commons path - Avoid unnecessary conversion (@@ -48,7 +48,7 @@) - ContainsOnlyLetter: this method will verify that a string contains only ASCII characters - Add more test
BugFix:
- RemoveFromString: will now check the index compared to string lenght (@@ -120,6 +132,9 @@)

+++ math/mathutils.go
@@ -48,7 +48,7 @@ func AverageFloat64(array []float64) float64 {
 	if array == nil {
 		return 0
 	} else if len(array) == 1 {
-		return float64(array[0])
+		return array[0]
 	}
 
 	for i := range array {
